Rozdzielenie komponentu Rooms na Room aby prościej zarządzało się stanem pojedynczej sali(usuwania, edycja) aktualnie wygląda to tak: komponent Rooms dostaje dane wszystkich pokoi i wyświetla liste komponentów Room komponent Room wyświetla dane pojedynczej sali lekcyjnej oraz odpowiada za edycje i usuwanie tej sali ze względu na dobrą strukturę takiego podejścia w przyszłości wszystkie komponenty będą pisane w taki sposób
